# PERF: avoid duplicate is_single_block check