Small bump, make up some of the ground lost in #34999.

```
In [2]: df = pd.DataFrame([1])                                                                                                                                                                                     

In [3]: %timeit df.values                                                                                                                                                                                          
3.65 µs ± 75.7 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <-- PR
3.84 µs ± 139 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <-- master
```